separate element searcher into InPageElement.ts
